Add native peer method for MethodHandles.privateLookupIn
